tools/swe-agent/requirements.txt
anthropic[bedrock]
config
datasets
docker >= 7.1.0  # https://github.com/princeton-nlp/SWE-agent/issues/379
gymnasium
numpy
openai>=1.0
pandas
rich
ruamel.yaml
swebench==2.0.13
tenacity
unidiff
simple-parsing
# Versions of together below 1.1.0 are not compatible. see https://github.com/princeton-nlp/SWE-agent/issues/135
together>=1.1.0
ollama
rich-argparse
flask
flask-cors
flask-socketio
bugswarm-common
